Moved ISelectionStrategy and ICrossoverStrategy to their designated directories

+++ inc/Services/GAConfig.h
@@ -2,8 +2,8 @@
 
 struct GAConfig
 {
-	int epochs{ 5 };
-	int population{ 50 };
+	size_t epochs{ 5 };
+	size_t population{ 50 };
 	int selectionMethod{ 0 };
 	int crossoverMethod{ 0 };
 	int fitnessMethod{ 0 };
